Separate external service config from legacy (#2306)
https://eaflood.atlassian.net/browse/WATER-4781

We're intending to switch from using the [notifications-node-client](https://github.com/alphagov/notifications-node-client) to using [GOV.UK Notify's REST API](https://docs.notifications.service.gov.uk/rest-api.html) directly.

> See [Switch to using the Notify REST API from package](https://github.com/DEFRA/water-abstraction-system/pull/2178) for more details about why.

To do this, we need to add **Notify's** REST API URL to the project. But this raised some questions!

- Do we add it to the existing `config/notify.config.js` along with the rest of the **Notify** config?
- Do we add it to the existing `config/services.config.js`, which is where all external API URLs currently live?
- If we do the later **Notify's** config will be split. So, should we ditch `config/notify.config.js`?

Essentially, this exposed that we have not been consistent with our config, so now we're not sure what to do.

This change resolves the issue with a re-organisation of our config. `config/services.config.js` will be renamed `config/legacy.config.js` and will be dedicated to the config needed to communicate with the legacy WRLS apps.

All other services will get their own config file.

Our view is that those browsing the project will get a clearer sense of what external APIs we are integrating with. Within the config files, we'll have more space and freedom to document the purpose of the API and the required configuration settings.