Add dynamic meter readings to the return forms PDF (#2299)
* Add dynamic meter readings to the return forms PDF

https://eaflood.atlassian.net/browse/WATER-5162

When generating meter readings, we need to handle them spilling over onto another page.

The meter reading uses boxes in the PDF form, this is presented as rows and columns.

This change calculates the number of dates we put on a single page; when this limit is reached, a new page will be added below the existing, and this continues.

We use the legacy codes CSS and some of the HTML to show the pages correctly, legacy code has stipulated the columns and rows as the expected format when rendering the PDF.