Add dynamic meter readings to the return forms PDF for days
https://eaflood.atlassian.net/browse/WATER-5162

We previously implemented the month and week meter readings.

This change adds the meter readings for the 'days' frequency.

When the return frequency is 'days', we show three columns (and set three columns in the HTML dues to the way the CSS works).

The day meter readings show the Month and Year at the top of each column, and each cell has the day and then the boxes. it looks something like this

June 2024
1 [][][][][][][][][][]
2 [][][][][][][][][][]

This differs from out established logic, so it has been implemented on its own within the presenter.

This change uses (and follows) the previously established calculation for the number of dates we put on a single page, and when this limit is reached, a new page will be added below the existing, and this continues.

We use the legacy codes CSS and some of the HTML to show the pages correctly, legacy code has stipulated the columns and rows as the expected format when rendering the PDF.