Add additional recipient - use route to save details (#2308)
https://eaflood.atlassian.net/browse/WATER-5047

As part of sending out notifications to users we need to be able to submit an address.

When the address has been submitted it is converted into a recipient object and added to the additionalRecipient array as well as the hash added to the selectedRecipient array.

This PR refactors the way the address is handled before being passed back to the check recipients page.