Add the return forms preview page
https://eaflood.atlassian.net/browse/WATER-5162

We previously had a placeholder to render just the first due return when the preview link was pressed.

This change adds a new page to allow the user to choose to preview any previously selected return log. This page follows a similar pattern to the check alerts preview page.

We show a table with some relevant data to the return log and a preview link. This preview link opens a new tab and sends a PDF. This will default to the browser PDF viewer.