Add new Notify Token Cache plugin
This copies the same idea of cache a generated JSON web token that we use for requests to the Charging Module.

Essentially, we are adding a hapi server method. They offer the ability to cache the return value. This will reduce our overhead when sending multiple requests to Notify, because we'll only need to generate the signed JSON web token once, rather than for each individual request.

We use a plugin to ensure the method is registered with Hapi when the server starts up.