Update the standard layout to provide the error summary, back link and page heading html. (#2320)
* Update the standard layout to provide the error summary, back link and page heading html.

To use the updated layout and keep the old pages working, we have introduced a new block called 'pageContent'. This should generally be used to show the form content of a page.

If a page needs to control the entire content, then the 'content' block can still be used. This also allows existing pages to still function as expected.

The backlink works on the same principle where the current pages will override the 'breadcrumbs' block. When the new 'pageContent' block is used, the override can be removed, but the backlink must now provide an object with the text and href.

The error messages will be provided by the submit logic for a page, and they will provide the 'errorList'.