Rename to CreateEmailRequest
Following the convention of the other requests, we do not normally start the module name with the name of the thing it connects to. That is for the folder name.

Admittedly, `SendEmailRequest` would be the closest match based on the Notify docs, but then we'd be calling `SendEmailRequest.send()` which also seemed a little off.

Where we cause a record to be 'created', we use the verb 'create', so that is what we have settled for here and and what we'll use for sending a letter.