Switch to using the Notify REST API from package (#2178)
https://eaflood.atlassian.net/browse/WATER-4781

In this last year, we have migrated the send return invitations and reminders notification journeys, which required sending requests to [GOV.UK Notify](https://www.notifications.service.gov.uk/) for the first time.

When we first started, we hit a blocker dealing with the proxy in our AWS environments. We had to resort to providing the [notifications-node-client](https://github.com/alphagov/notifications-node-client) with a customised instance of [Axios](https://github.com/axios/axios) to get it to work.

However, we then needed to add more request types. Later still, we added some requests to the [ea-address-facade](https://github.com/DEFRA/ea-address-facade). This address work highlighted that our Notify 'requests' were very different from our other requests. We also had a long-standing issue with **Snyk** flagging a vulnerability with **Axios** that the Notify package was using (though we later [figured out a fix](https://github.com/DEFRA/water-abstraction-system/pull/2248)).

This led us to re-examine the ways we could integrate with **Notify**. We realised that if we switched to using their REST API instead of the client that

- We could drop the package, including its dependence on **Axios** (an alternate to [Got](https://github.com/sindresorhus/got)
- We could drop our workaround for dealing with the proxy in our AWS environments
- We could build on the work we'd already invested into our `base.request.js` module, with its ability to retry specific failed requests automatically
- _All_ the Notify modules could be updated to follow our conventions for 'requests', making them consistent and easier to maintain
- We might even get a slight performance bump when we are sending multiple requests. The client package generates a JSON web token for each request. However, the token has a 30-second expiry, so it could be reused.

This changes switches our **GOV.UK Notify** integration to use their REST API instead of the npm package.