Add Notify service to our health info page (#2330)
https://eaflood.atlassian.net/browse/WATER-4781

We now integrate with [GOV.UK Notify](https://www.notifications.service.gov.uk/) for sending emails and letters. However, even though this is live, we overlooked adding the service to our `/health/info` page.

This adds it so we can confirm we can connect to the service and it is up and running from our info page.