Housekeeping - Be explicit about responseType
Having to set the responseType for the Gotenberg request made us think about how we were doing it in the Address Facade version. Though Got defaults to 'text' we decided it would be better to make that explicit in the request. These request objects are a way of declaring exactly how to send _that_ request to _that_ service, so we should be verbose.