Refactor Gotenberg ViewHealthRequest to use Base  (#2331)
https://github.com/DEFRA/water-abstraction-team/issues/150

In [Refactor Gotenberg health check to status request](https://github.com/DEFRA/water-abstraction-system/pull/2325), we moved the request that was being made in `InfoService` directly to our [Gotenberg](https://gotenberg.dev/) to its own 'request'.

As we worked through the other apps `InfoService` checks, we noticed an inconsistency. **Gotenberg** was having to parse the result into JSON, whereas others could refer to it directly.

That's when we spotted that `GotenbergRequest` assumes all requests want the [responseType](https://github.com/sindresorhus/got/blob/main/documentation/2-options.md#responsetype) set to `'buffer'`.

This is valid in the sense we are using it to convert HTML to PDF, and **Gotenberg** sends the PDF data back as form data. But in the case of just checking its health endpoint, JSON is what we want.

So, we'll use the same pattern we did for the [Address Facade](https://github.com/DEFRA/water-abstraction-system/pull/2326) (its health endpoint returns plain text, whereas all other endpoints return JSON). We'll update the **Gotenberg** `ViewHealthRequest` to use `BaseRequest` instead.