Delete redundant /check endpoints
https://eaflood.atlassian.net/browse/WATER-5147

In [Add /check end points to verify Notify address fix](https://github.com/DEFRA/water-abstraction-system/pull/2193) we added some `/check` endpoints to allow us to verify the addresses we were generating when sending letters to Notify would better match what it requires.

That actual change to how we send addresses is now implemented and live, and being used as part of sending notices on a daily basis.

This means these `/check` endpoints are now redundant and can be deleted.