Add replacement route and update controller
The route, controller and job should always align. We can't get rid of the old route just yet because we need this change to drop into production. We can then update the Jenkins pipeline that triggers the job to use the 'fixed' route without breaking anything, at which point we can delete the old route.

For now though, both point to the same 'fixed' controller name.