Add standard logging to notifications status job (#2333)
https://eaflood.atlassian.net/browse/WATER-4961

When we added the [notifications status job](https://github.com/DEFRA/water-abstraction-system/pull/1842), we appear to have omitted the usual logging that we perform for all jobs. Usually, we get confirmation that the job has completed, how long it took, and an idea of how much work was done.

We also didn't follow our standard naming conventions for the service.

This change rectifies those omissions.