Remove use of title filter in return forms PDF
https://eaflood.atlassian.net/browse/WATER-5162

It was spotted during testing that the PDF return forms we are generating were not displaying a returns site description as shown in the UI.

When we looked we found that the PDF view was using [Nunjucks title() filter](https://mozilla.github.io/nunjucks/templating.html#title). Admittedly, we got the HTML from the legacy, but that does not make this behaviour right.

For both the purpose and the site description we should be including them in the form 'as is'.

This removes the filter from the cover view.