Fix issues with FetchNotice inc. unsafe SQL injection
The unsafe SQL injection was the first thing that needed to be fixed. SQL queries should _never_ use string interpolation to build a query as this exposes you to [SQL injection attacks](https://owasp.org/Top10/A03_2021-Injection/).

After that we have tidied up the naming conventions (`notice > event` and `notifications > results`), corrected some typos, added the notify status to the query (needed if we want to provide a more meaningful update), and tweaked the query to avoid having to mix snake and camelcase.