Update FetchNotice tests plus corrections
We had to update the tests because of the changes we made to the fetch service. But we also ensured that we created both a letter and an email notification, and set one as errored.

On top of that we corrected the inconsistencies (mixing abstraction alert and return reminder values), as well as put the right statues in the right records ('completed' goes in the event and 'sent' goes in the notification).