Update FetchNotice to use Objection
By using Objection it will mean we can then plug in our pagination into the query. Prior to this change, we were always fetching _all_ notifications and then slice what to display in code.

Now we'll be able to only query for the notifications we displaying, improving the performance and reducing the amount of code we have to maintain.