Add data-test tags for acceptance tests (#2339)
https://eaflood.atlassian.net/browse/WATER-4892

As part of the work refactoring the acceptance tests for the return logs and return version journeys it was noticed that we're not checking the summary page. Adding that page onto the end of some of the journeys requires some additional data-test tags so the acceptance tests can correctly identify which areas to verify.