Returns tab - No due date https://eaflood.atlassian.net/browse/WATER-5213
Currently when a return log is created it is automatically pre-populated with a due date. However, the process of assigning a due date to a return log is being updated so that the due date is not set on creation. Instead, it is set when the reminders are sent out.

This means that some new return logs will now have a null due date, where previously the due date would always have been populated.

So in this PR the Returns tab of a licence is going to be updated to show a blank due date when it is null. As well as the status of "Not due yet".