Update download recipients to handle additional and selected recipients (#2322)
https://eaflood.atlassian.net/browse/WATER-5223

The logic to handle the recipients has become quite complicated. We need to simplify the handling recipients.

We now have additional and selected recipients we need to merge into the recipients array. We need to do this in multiple places.

This change adds a new service to handle merging our recipients form the database and those manually added by the user (additional recipients). This will be the common place to get the recipients.

As part of this change we have refactor some of our service names to align with the new change.

This change also fixes the download by using this new'RecipientsService' to add the additional recipients and selected recipients to the CSV.