Add request service to send a precompiled file to Notify (#2344)
* Add request service to send a precompiled file to Notify

We have made changes to incorporate the legacy PDF generation notices into this repo.

We can create the PDF, now we need to send it to Notify.

This change adds the CreatePrecompiledFileRequest' code to send a file to notify. This file needs to be a valid PDF as mentioned in the documentation. This has been linked in the docs above the request logic.

When a base64 encoded string is provided, it will send the file to notify; when a request fails, it will fall into our existing error flow.