Update return forms services in preparation for sending notices (#2345)
* Update return forms services in preparation for sending notices

https://eaflood.atlassian.net/browse/WATER-5226

We had some assumptions on how the return forms services would work. We need to have the calling services control the recipient and the due return log.

This change updates the preview return forms services to get the due return log selected previously by the user.

This change will enable another service (not yet implemented) to iterate over the selected due return logs and recipients and send a notice to each recipient of the due return log.

This should simplify the 'PrepareReturnFormsService' to only be concerned with creating (and now returning) the page data and the file.