Add recipient download specific variables to additional contacts
https://eaflood.atlassian.net/browse/WATER-5215

We have added a change to allow the user to add additional recipients. We have implemented the logic to render and send these additional recipients a notice.

Theses additional recipients are shown in the download CSV. But they are missing the 'Licence' and 'Contact Type' fields (they are blank).

This change adds the relevant fields when the additional licence is saved.

This allows the existing logic to remain the same and still open to extension.

This does result in us setting the licence ref twice in the additional recipient. Once on 'licence_refs' and then on 'licence_ref'. This is because the rendering and sending of recipients share the same logic with the other journeys which can have multiple licence refs per recipients.

We set the 'contact_type' as 'Single use' for the download only. This will only be rendered in the CSV.

We have the 'DetermineRecipientsService' which has logic to check the contact types, if these prior conditions are not met, then the default will be 'Returns agent' (for an email) and 'Returns to' (for a letter).

This implied logic has been highlighted by creating the '_addDownloadRecipientData' function. This has JSDocs added to further illustrate the implied logic.