Update notifications 4 auto-populated created_at (#2347)
https://eaflood.atlassian.net/browse/WATER-5229

The legacy code always populates the `send_after` field with the current time when creating a notification, and only sometimes the `date_created` field.

The new code in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) always populates `date_created`, and never populates `send_after`. We now consider it a redundant field.

Our issue is that our acceptance tests rely on an endpoint that returns the last notification created for a specified email address. But the disparity between how the two systems create notifications means it can no longer find the last email, which has broken some of the tests.

We need to fix it so it can!

In [Auto-populate scheduled_notification date_created](https://github.com/DEFRA/water-abstraction-service/pull/2716) we added a migration to

- update the `water.scheduled_notification` table so that `date_created` will now automatically populate when a new record is inserted
- update existing records setting `date_created` to `send_after` where `date_created` is null and `send_after` isn't

Now that change has been made, we need to update our legacy migrations accordingly. We can also eliminate the need to manually set the value in our helper.