Remove 'allRecipients' flag from recipients service (#2350)
https://eaflood.atlassian.net/browse/WATER-5233

We previously implemented an 'allRecipients' flag for the recipients service to know to render all the recipients. This was only needed for the selected recipients (and the check and downloaded service subsequently need to set this).

This change removes the 'allRecipients' flag.

We now just temporarily remove the 'session.selectedRecipients' when we pass it into the session and pass in the selected recipients as an additional arg.

This allows the view to take a copy of the 'selectedRecipients' and pass it in, and the submit service to pass in the user's payload of selected recipients.