Scaffold notice setup recipient name (#2351)
When sending a paper form, we want to allow the user to send the form to an additional recipient.

Our existing page offers a choice between email and post (a letter). We do not want to show the email option to the user.

The simplest way we can do this is by adding a standalone page to enter the recipient name. This will redirect to the generic address service the same as the existing screen.

This change adds the scaffolded code for the recipient name page, the view and submit.