View return log - fix status when no due date (#2349)
https://eaflood.atlassian.net/browse/WATER-5214

Currently when a return log is created it is automatically pre-populated with a due date. However, the process of assigning a due date to a return log is being updated so that the due date is not set on creation. Instead, it is set when the reminders are sent out.

This means that some new return logs will now have a null due date where previously the due date would always have been populated.

At the moment when viewing a return log that has no due date the status is incorrectly shown as "Overdue", when it should actually be "Not due yet".

This PR will update how the status is determined when there is no due date.