Housekeeping - Move pageTitle to presenter
Both the view and the submit service have the same code to determine the pageTitle. By moving it to the presenter we remove the duplication.

To do this we have to pass in the information about selected page and number of pages. The view also needs this info, so we have the presenter return it, but switch to more generic names to help with consistency when other pages wish to follow the same pattern.