Switch to pageContent and strip out redundant stuff
We update the view to override the `pageContent` block rather than the parent `content` block.

This means we can make use of the error summary and pageHeadingHtml from the main `layout.njk`.

We strip out all the extra div's we now know were not doing anything.

We also go through and re-order the properties to be alphabetical.