Remove the backlink on the notice setup check page when the journey is 'adhoc'
https://eaflood.atlassian.net/browse/WATER-5234

We are seeing a few journey issues for the adhoc journey when a goes back from the check recipients page.

We have an established pattern of not having backlinks on the check page. We have the backlink on this page for the invitation and reminders journey, as they do not have their own check pages.

This change conditionally renders the backlink based on the journey.

This change also updates the nunjucks to use our new layout.