Update flash notification to support the updated layout
https://eaflood.atlassian.net/browse/WATER-5234

We have updated the layout to include the error summary and backlink.

We can also provide the notification banner.

This change adds the notification banner to the updated layout.

We are updating the way we use the notifications by having the layout automatically provide the GDS Notification
banner macro.

This is how we provide the notification macro:
```
{% if notification %}
  {{ govukNotificationBanner(notification) }}
{% endif %}
```

We are in the process of migrating pages to use this improved layout. But we need to support the existing way of
presenting a notification. This is how we do it:
```html
{{ govukNotificationBanner({
     titleText: notification.title,
     text: notification.text
}) }}
```

So we need to support two ways of presenting a notification. The updated layout expects the notification object to
have 'titleText' and the 'text' to be present to render the notification, the existing process expects the 'title'
and 'text'.

Fortunately, we can set both without the value conflicting.