Move countries logic into a base presenter
We are aware that we only use this method once in the actual code. But in order to test the presenter, service and submit service we've had to duplicate it in 3 other places.

So, we move it out of `InternationalPresenter` in a `BaseAddressPresenter`, so that we can reuse the same logic in both the presenter and the various tests.