Set due date to null when creating return logs (#2352)
https://eaflood.atlassian.net/browse/WATER-5212

To accommodate new legislation we need to track when a customer has been notified to submit their return logs. To do that we need to not set the due date when we create the return logs.

This PR adds in the logic to create a return log and leave the due date null. This will be set once we have confirmed that the user has been successfully notified of their returns.