Returns tab - Fix incorrect status (#2353)
https://eaflood.atlassian.net/browse/WATER-5213

During QA of the work done in https://github.com/DEFRA/water-abstraction-system/pull/2346 it has been spotted that when the status is "void", and the due date is `null`. The status is incorrectly being displayed as "Not due yet".

This PR wil fix that issue by using the base presenters `formatReturnLogStatus` rather than the current private function.

`formatReturnLogStatus` has been moved from the `view-licence-returns.presenter.js` where it was previously called `formatStatus`.