Add new SubmitCheckNoticeTypeService
Previously, the page's green button when clicked just took the user to the next page in the journey. However, this is the point we now want to set the `addressJourney`, knowing we need to configure it for the different notice types.

So, we'll need to change the page to use a form, which will submit a POST request. This is the service where that request will end up, which will perform the work of updating the session with the `addressJourney`.