Update the contact type page to show the page title caption
https://eaflood.atlassian.net/browse/WATER-5236

We have missed adding the caption to the contact type page.

This change adds the page caption to the contact types page.

We have updated our base layout to provide the error summary and backlink. This change has updated the content block to use this improved layout.

This change also uses the common validator.