Prep for supporting a title caption
Another thing we have realised our generic address journey doesn't support but we'd like it to for ad-hoc is a 'caption'. We frequently use these in our pages to tie them together.

Captions are nicely handled in our new "template standad", and with the changes we've already done, we are so close we might as well complete the job.

So, the next bout of commits are changes help get the pages to template standard. We'll then add what we need to support a caption in the address pages.