Refactor expected session object for address lookup (#2362)
https://eaflood.atlassian.net/browse/WATER-5044

As part of the setup ad-hoc return invitation and return form notice journeys (that's a mouthful!), we need to provide users with the ability to add additional recipients. This includes being able to enter an address.

There are multiple journeys in the service that require an address to be entered, so we planned from the start to make our address setup pages 'generic', i.e. they could be used as part of other setup journeys.

Essentially, there are some things we need to tell the address journey.

- Where to go back to if a user clicks 'Back' on the first page
- Where to redirect to when an address has been successfully submitted

There are also some things the address journey needs to provide to the parent journey.

- the address the user entered

Using the notice setup journey as our guinea pig, we had an initial idea of using `session.address` and setting it up in its `initiate.service.js`. But a recent change to the journey has highlighted that's not quite right.

- We forgot about telling the address journey which back link to use!
- We forgot about telling the address journey which nav bar to show as active!
- We forgot about adding support for displaying a caption in the address journey
- We don't know what back and redirect URLs to use until later in the journey, so `initiate.service.js` is the wrong place (though it might be fine in future journeys)
- Because everything is a top-level property of `session.address`, how we reset it for reuse now feels overly complex.

So, we're going to change a few things.

For general use, we will configure the address journey to expect the following in the session.

```javascript
session.addressJourney = {
  activeNavBar: 'manage',
  address: {},
  backLink: { href: '/system/notices/setup/foo', text: 'Back' },
  pageTitleCaption: `Notice ${session.referenceCode}`, // Optional
  redirectUrl: `/system/notices/setup/${session.id}/add-recipient`
}
```

This will also better support the ad-hoc notice setup journey. Specific to it, we allow users to go through the address journey multiple times (most journeys will go through it once). So, it needs to 'reset' the session so as not to replay previously entered address details. Holding them separately from the config should make that process cleaner.

In this change, we'll also move when we apply the `addressJourney` object to the session, to the point we know what additional recipient journey is needed, and after [Remove the backlink on the notice setup check page when the journey is 'adhoc'](https://github.com/DEFRA/water-abstraction-system/pull/2361) is merged, it will be 'locked in'.

---

Because of the number of changes that needed to be made to the address pages, we also ended up applying our "template standard"; setting `pageContent` in our views and letting `layout.njk` control things like the page title, caption, and error summary.